Patches bug that causes omission of groups.
This commit patches a bug causing the omission of multiple protected
group identities from the multicalibration process, both during training
and evaluation. The authors thank Dutch Hansen for finding this issue. This
commit patches this bug and doubles the default number of training iterations
used. This commit also includes some minor quality-of-life updates.